requirements.txt
torch>=1.7.0
torchaudio>=0.7.0
numpy
scipy
scikit-learn
pyyaml
soundfile

--extra-index-url https://pypi.nvidia.com
cuml-cu12==24.8.*